requirements.txt
absl-py
django>=2.1.6